pyproject.toml
[project]
name = "zhiqiao-c4"
version = "0.1.0"
description = "知桥 C4 - 多作者内容追踪系统"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "feedparser>=6.0.0",
    "lxml>=5.0.0",
    "python-dateutil>=2.8.0",
    "flask>=3.0.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
]
